% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{coerce-numeric-sparse}
\alias{coerce-numeric-sparse}
\title{Coerce Numeric Vector to sparse_numeric}
\arguments{
\item{from}{A numeric vector.}
}
\value{
A sparse_numeric object.
}
\description{
Converts a numeric vector into a sparse_numeric representation.
}
\keyword{internal}
