% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{coerce-sparse-numeric}
\alias{coerce-sparse-numeric}
\title{Coerce sparse_numeric to Numeric Vector}
\arguments{
\item{from}{A sparse_numeric object.}
}
\value{
A numeric vector.
}
\description{
Converts a sparse_numeric object into a full numeric vector.
}
\keyword{internal}
