% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{norm}
\alias{norm}
\alias{norm,sparse_numeric-method}
\title{Squared Norm of a sparse_numeric Vector}
\usage{
norm(x, ...)
}
\arguments{
\item{x}{A sparse_numeric object.}

\item{...}{Additional arguments (unused).}
}
\value{
A numeric scalar.
}
\description{
Computes the Euclidean norm (L2 norm) of the sparse vector.
}
