% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{sparse_add}
\alias{sparse_add}
\alias{sparse_add,sparse_numeric,sparse_numeric-method}
\title{Add Two Sparse Numeric Vectors}
\usage{
sparse_add(x, y, ...)
}
\arguments{
\item{x}{A sparse_numeric object.}

\item{y}{A sparse_numeric object.}

\item{...}{Additional arguments (unused).}
}
\value{
A sparse_numeric object containing the sum.
}
\description{
Performs element-wise addition of two sparse_numeric vectors.
}
\seealso{
\code{\link{sparse_sub}}, \code{\link{sparse_mult}}
}
