% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{sparse_crossprod}
\alias{sparse_crossprod}
\alias{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\title{Cross Product of Two Sparse Numeric Vectors}
\usage{
sparse_crossprod(x, y, ...)
}
\arguments{
\item{x}{A sparse_numeric object.}

\item{y}{A sparse_numeric object.}

\item{...}{Additional arguments (unused).}
}
\value{
A numeric scalar.
}
\description{
Computes the scalar inner product (cross-product) of two sparse_numeric vectors.
}
\seealso{
\code{\link{sparse_mult}}
}
