% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{sparse_mult}
\alias{sparse_mult}
\alias{sparse_mult,sparse_numeric,sparse_numeric-method}
\title{Multiply Two Sparse Numeric Vectors}
\usage{
sparse_mult(x, y, ...)
}
\arguments{
\item{x}{A sparse_numeric object.}

\item{y}{A sparse_numeric object.}

\item{...}{Additional arguments (unused).}
}
\value{
A sparse_numeric object containing the element-wise product.
}
\description{
Performs element-wise multiplication of two sparse_numeric vectors.
}
\seealso{
\code{\link{sparse_add}}, \code{\link{sparse_sub}}
}
