% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\docType{class}
\name{sparse_numeric-class}
\alias{sparse_numeric-class}
\title{sparse_numeric S4 Class}
\description{
Represents a sparse numeric vector using only non-zero elements.
}
\section{Slots}{

\describe{
\item{\code{value}}{A numeric vector containing non-zero values.}

\item{\code{pos}}{An integer vector giving the positions of non-zero values.}

\item{\code{length}}{An integer giving the full vector length.}
}}

