% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{sparse_numeric-plot}
\alias{sparse_numeric-plot}
\alias{plot,sparse_numeric,sparse_numeric-method}
\title{Plot Overlapping Non-Zero Elements of Two Sparse Vectors}
\usage{
\S4method{plot}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{A sparse_numeric object.}

\item{y}{A sparse_numeric object.}

\item{...}{Additional graphical arguments passed to \code{plot()}.}
}
\value{
A scatterplot.
}
\description{
Visualizes the positions and values where two sparse_numeric vectors both
have non-zero entries.
}
