% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{sparse_numeric-show}
\alias{sparse_numeric-show}
\alias{show,sparse_numeric-method}
\title{Show Method for sparse_numeric}
\usage{
\S4method{show}{sparse_numeric}(object)
}
\arguments{
\item{object}{A sparse_numeric object.}
}
\description{
Displays a sparse_numeric object in a human-readable summary.
}
