% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{sparse_sub}
\alias{sparse_sub}
\alias{sparse_sub,sparse_numeric,sparse_numeric-method}
\title{Subtract Two Sparse Numeric Vectors}
\usage{
sparse_sub(x, y, ...)
}
\arguments{
\item{x}{A sparse_numeric object.}

\item{y}{A sparse_numeric object.}

\item{...}{Additional arguments (unused).}
}
\value{
A sparse_numeric object containing the difference.
}
\description{
Performs element-wise subtraction of two sparse_numeric vectors.
}
\seealso{
\code{\link{sparse_add}}, \code{\link{sparse_mult}}
}
