% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{standardize}
\alias{standardize}
\alias{standardize,sparse_numeric-method}
\title{Standardize a sparse_numeric Vector}
\usage{
standardize(x, ...)
}
\arguments{
\item{x}{A sparse_numeric object.}

\item{...}{Additional arguments (unused).}
}
\value{
A sparse_numeric object.
}
\description{
Produces a standardized sparse_numeric vector with mean zero and unit variance.
}
